requirements.txt
matplotlib>=3.3.0
pydicom>=2.0.0
pyamg>=4.0.0
scipy>=1.4.1
numpy>=1.18.5
scikit_image>=0.17.2
Pillow>=7.2.0
PyQt5>=5.15.0
qimage2ndarray>=1.8.3
